Fix discovery logic in OnAffinityTopologyVersionChange
