+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Client/Cache/PartitionAwarenessTest.cs
@@ -22,7 +22,6 @@ namespace Apache.Ignite.Core.Tests.Client.Cache
     using System.Globalization;
     using System.Linq;
     using System.Net;
-    using System.Runtime.CompilerServices;
     using System.Threading.Tasks;
     using Apache.Ignite.Core.Cache.Affinity.Rendezvous;
     using Apache.Ignite.Core.Cache.Configuration;
@@ -41,7 +40,7 @@ namespace Apache.Ignite.Core.Tests.Client.Cache
         private const int ServerCount = 3;
 
         /** */
-        private ICacheClient<int, int> _cache;
+        protected ICacheClient<int, int> _cache;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PartitionAwarenessTest"/> class.
@@ -151,44 +150,6 @@ namespace Apache.Ignite.Core.Tests.Client.Cache
 
             Assert.AreEqual(expected, ex.Message);
         }
-
-#if NETCOREAPP // TODO: IGNITE-15710
-        [Test]
-        public void CacheGet_NewNodeEnteredTopology_RequestIsRoutedToNewNode()
-        {
-            // Warm-up.
-            Assert.AreEqual(1, _cache.Get(1));
-
-            // Before topology change.
-            Assert.AreEqual(12, _cache.Get(12));
-            Assert.AreEqual(1, GetClientRequestGridIndex());
-
-            Assert.AreEqual(14, _cache.Get(14));
-            Assert.AreEqual(2, GetClientRequestGridIndex());
-
-            // After topology change.
-            var cfg = GetIgniteConfiguration();
-            cfg.AutoGenerateIgniteInstanceName = true;
-
-            using (Ignition.Start(cfg))
-            {
-                TestUtils.WaitForTrueCondition(() =>
-                {
-                    // Keys 12 and 14 belong to a new node now (-1).
-                    Assert.AreEqual(12, _cache.Get(12));
-                    if (GetClientRequestGridIndex() != -1)
-                    {
-                        return false;
-                    }
-
-                    Assert.AreEqual(14, _cache.Get(14));
-                    Assert.AreEqual(-1, GetClientRequestGridIndex());
-
-                    return true;
-                }, 6000);
-            }
-        }
-#endif
 
         [Test]
         [TestCase(1, 1)]
@@ -609,7 +570,7 @@ namespace Apache.Ignite.Core.Tests.Client.Cache
             return cfg;
         }
 
-        private int GetClientRequestGridIndex(string message = null, string prefix = null)
+        protected int GetClientRequestGridIndex(string message = null, string prefix = null)
         {
             message = message ?? "Get";
